backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pydantic[email]==2.5.2
email-validator==2.1.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
crewai==0.1.32
python-multipart==0.0.6
anthropic>=0.39.0
